Revert "[LER-91] Create Program Cards"
