+++ package.json
@@ -19,14 +19,12 @@
     "@types/react-dom": "^18.2.18",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "formik": "^2.4.5",
-    "polished": "^4.3.1",
     "react": "^18.2.0",
     "react-circular-progressbar": "^2.1.0",
     "react-dom": "^18.2.0",
     "react-redux": "^9.0.4",
     "react-router-dom": "^6.21.1",
     "react-scripts": "5.0.1",
-    "react-tooltip": "^5.26.3",
     "styled-components": "^6.1.6",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4",
