--- src/assets/icons/QuestionnaireIcon.tsx
@@ -1,19 +1,0 @@
-import { IconInterface } from '../../utils/utils';
-import React from 'react';
-
-export const QuestionnaireIcon = ({ size, color }: IconInterface) => {
-  return (
-    <svg
-      width={size ?? '18'}
-      height={size ?? '18'}
-      viewBox="0 0 18 18"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M4.89047 17.4709C4.62873 17.4709 4.4239 17.3856 4.27596 17.2149C4.12802 17.0442 4.05405 16.8166 4.05405 16.5321V14.2106H3.79801C3.04694 14.2106 2.38691 14.0541 1.81792 13.7412C1.24893 13.4225 0.802276 12.973 0.477952 12.3927C0.159317 11.8066 0 11.1067 0 10.2931V3.92609C0 3.11244 0.156472 2.41258 0.469417 1.82652C0.782361 1.24046 1.23186 0.790957 1.81792 0.478013C2.40398 0.159378 3.10384 6.10352e-05 3.9175 6.10352e-05H14.0825C14.9018 6.10352e-05 15.6046 0.159378 16.1906 0.478013C16.7767 0.796647 17.2233 1.24899 17.5306 1.83505C17.8435 2.41542 18 3.11244 18 3.92609V10.2931C18 11.1067 17.8435 11.8066 17.5306 12.3927C17.2233 12.973 16.7767 13.4225 16.1906 13.7412C15.6046 14.0541 14.9018 14.2106 14.0825 14.2106H9.0128L5.96586 16.9247C5.74964 17.1181 5.56188 17.2575 5.40256 17.3429C5.24324 17.4282 5.07255 17.4709 4.89047 17.4709ZM8.79943 8.70561C9.06117 8.70561 9.27454 8.64586 9.43954 8.52638C9.60455 8.40689 9.70128 8.25326 9.72973 8.06549C9.73542 8.04842 9.73826 8.03135 9.73826 8.01429C9.73826 7.99722 9.73826 7.98299 9.73826 7.97161C9.7724 7.77815 9.86629 7.60461 10.0199 7.45099C10.1792 7.29736 10.3812 7.14089 10.6259 6.98157C11.0583 6.70276 11.3883 6.42111 11.6159 6.13662C11.8492 5.85212 11.9659 5.47375 11.9659 5.00148C11.9659 4.52922 11.835 4.12239 11.5733 3.781C11.3172 3.43961 10.9701 3.17787 10.532 2.99579C10.0939 2.80803 9.6074 2.71414 9.07255 2.71414C8.34993 2.71414 7.74395 2.85355 7.25462 3.13235C6.77098 3.40547 6.4495 3.74686 6.29018 4.15653C6.26174 4.2305 6.23898 4.30732 6.22191 4.38697C6.20484 4.46663 6.1963 4.54914 6.1963 4.63448C6.1963 4.85639 6.27312 5.04131 6.42674 5.18925C6.58037 5.3315 6.76529 5.40262 6.98151 5.40262C7.1522 5.40262 7.30014 5.37133 7.42532 5.30874C7.5505 5.24615 7.66145 5.15227 7.75818 5.02709L7.8606 4.89053C7.99715 4.68 8.15078 4.52638 8.32148 4.42965C8.49787 4.33292 8.69132 4.28456 8.90185 4.28456C9.18634 4.28456 9.41679 4.36421 9.59317 4.52353C9.77525 4.67716 9.86629 4.87346 9.86629 5.11244C9.86629 5.3571 9.77809 5.55909 9.60171 5.71841C9.43101 5.87204 9.17496 6.06834 8.83357 6.30732C8.53769 6.51784 8.29303 6.73975 8.09957 6.97303C7.91181 7.20063 7.81792 7.48513 7.81792 7.82652C7.81792 7.8379 7.81792 7.84928 7.81792 7.86066C7.81792 7.87204 7.81792 7.88342 7.81792 7.8948C7.81792 8.43534 8.14509 8.70561 8.79943 8.70561ZM8.79943 11.4026C9.10669 11.4026 9.36558 11.3116 9.5761 11.1295C9.79232 10.9474 9.90043 10.7056 9.90043 10.404C9.90043 10.1082 9.79516 9.86919 9.58464 9.68712C9.37411 9.49935 9.11238 9.40547 8.79943 9.40547C8.48649 9.40547 8.22191 9.49935 8.00569 9.68712C7.78947 9.86919 7.68137 10.1082 7.68137 10.404C7.68137 10.7056 7.78947 10.9474 8.00569 11.1295C8.22191 11.3116 8.48649 11.4026 8.79943 11.4026Z"
-        fill={color ?? '#001B23'}
-      />
-    </svg>
-  );
-};
