+++ src/components/program/ProgramContent/index.tsx
@@ -32,7 +32,6 @@ const ProgramContent = () => {
           css={{
             width: '114px',
             height: '30px',
-            cursor: 'pointer',
           }}
         >
           Agregar pildora
